v.0.20.0 Trident integration

+++ example-stop.sh
@@ -1,4 +1,7 @@
 #!/bin/sh
 PORT=3000
-PID=`lsof -i4 | grep $PORT | grep node | awk '{print $2}'`
-kill $PID
+PID=`lsof -i4 | grep $PORT | grep node | awk '{print $2}')`
+if [ -n "$PID" ]; then
+	kill $PID
+	echo "OK"
+fi
